embed `baseStruct`
