+++ pkg/models/custom_timeline_date.go
@@ -1,0 +1,30 @@
+package models
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+)
+
+type CustomTimelineDateType string
+
+const (
+	CustomTimelineDateTypeSingle CustomTimelineDateType = "SINGLE"
+	CustomTimelineDateTypeRange  CustomTimelineDateType = "RANGE"
+)
+
+var AllCustomTimelineDateType = []CustomTimelineDateType{
+	CustomTimelineDateTypeSingle,
+	CustomTimelineDateTypeRange,
+}
+
+type CustomTimelineDate struct {
+	baseStruct
+
+	ID          uuid.UUID              `json:"id"`
+	Title       string                 `json:"title"`
+	Description *string                `json:"description,omitempty"`
+	DateType    CustomTimelineDateType `json:"dateType"`
+	StartDate   time.Time              `json:"startDate"`
+	EndDate     *time.Time             `json:"endDate,omitempty"`
+}
